+++ CHANGELOG.md
@@ -912,3 +912,4 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 * Fix custom dynamic variables not being colored in the Armaments preview (kiooeht)
 * debug console command can now be used without arguments to toggle debug mode (kiooeht)
 * info console command can now be used with arguments to set true/false (kiooeht)
+* Add evalcode console command for running code via the console (kiooeht)